Better naming convention for scalign scripts
